Cast sizeof result to int to fix test

+++ tests/boards/Generic/_005-LineByLine/libraries/LineByLine/LineByLine.cpp
@@ -9,7 +9,7 @@ LineByLine::LineByLine(char **pDataList)
 
 bool LineByLine::hasNext()
 {
-	return pos < sizeof(dataList);
+	return pos < (int) sizeof(dataList);
 }
 
 char* LineByLine::next()
